Combine error models into single file
