+++ packages/core/src/models/data/ModuleInfo.ts
