--- packages/core/src/models/data/error/ErrorInfo.ts
@@ -1,6 +1,0 @@
-import { InnerErrorInfo } from "./InnerErrorInfo.js";
-import { ModuleInfo } from "./ModuleInfo.js";
-
-export class ErrorInfo extends InnerErrorInfo {
-  modules?: ModuleInfo[];
-}
