--- packages/core/src/models/data/error/MethodInfo.ts
@@ -1,12 +1,0 @@
-import { ParameterInfo } from "./ParameterInfo.js";
-
-export class MethodInfo {
-  data?: any;
-  generic_arguments?: string[];
-  parameters?: ParameterInfo[];
-  is_signature_target?: boolean;
-  declaring_namespace?: string;
-  declaring_type?: string;
-  name?: string;
-  module_id?: number;
-}
