--- packages/core/src/models/data/error/ParameterInfo.ts
@@ -1,7 +1,0 @@
-export class ParameterInfo {
-  data?: any;
-  generic_arguments?: string[];
-  name?: string;
-  type?: string;
-  type_namespace?: string;
-}
