--- packages/core/src/models/data/error/StackFrameInfo.ts
@@ -1,7 +1,0 @@
-import { MethodInfo } from "./MethodInfo.js";
-
-export class StackFrameInfo extends MethodInfo {
-  file_name?: string;
-  line_number?: number;
-  column?: number;
-}
