+++ packages/core/src/services/IErrorParser.ts
@@ -1,4 +1,4 @@
-import { ErrorInfo } from "../models/data/error/ErrorInfo.js";
+import { ErrorInfo } from "../models/data/ErrorInfo.js";
 import { EventPluginContext } from "../plugins/EventPluginContext.js";
 
 export interface IErrorParser {
